requirements.txt
# Python dependencies
pandas
scikit-learn
fastapi
uvicorn
pytest
pyyaml
joblib
